Encapsulated RedisActor in connector module - increased privacy restriction
